Refactor unix window backends
