+++ src/SFML/Window/CursorImpl.hpp
@@ -36,7 +36,7 @@
 
 #elif defined(SFML_SYSTEM_LINUX) || defined(SFML_SYSTEM_FREEBSD) || defined(SFML_SYSTEM_OPENBSD) || defined(SFML_SYSTEM_NETBSD)
 
-    #include <SFML/Window/Unix/CursorImpl.hpp>
+    #include <SFML/Window/X11/CursorImpl.hpp>
 
 #elif defined(SFML_SYSTEM_MACOS)
