+++ src/SFML/Window/Generic/SensorImpl.cpp
