+++ src/SFML/Window/WindowImpl.cpp
@@ -43,10 +43,10 @@
 
 #elif defined(SFML_SYSTEM_LINUX) || defined(SFML_SYSTEM_FREEBSD) || defined(SFML_SYSTEM_OPENBSD) || defined(SFML_SYSTEM_NETBSD)
 
-    #include <SFML/Window/Unix/WindowImplX11.hpp>
+    #include <SFML/Window/X11/WindowImplX11.hpp>
     typedef sf::priv::WindowImplX11 WindowImplType;
 
-    #include <SFML/Window/Unix/VulkanImplX11.hpp>
+    #include <SFML/Window/X11/VulkanImplX11.hpp>
     typedef sf::priv::VulkanImplX11 VulkanImplType;
 
 #elif defined(SFML_SYSTEM_MACOS)
