+++ src/SFML/Window/X11/ClipboardImpl.hpp
