+++ src/SFML/Window/X11/VideoModeImpl.cpp
@@ -26,7 +26,7 @@
 // Headers
 ////////////////////////////////////////////////////////////
 #include <SFML/Window/VideoModeImpl.hpp>
-#include <SFML/Window/Unix/Display.hpp>
+#include <SFML/Window/X11/Display.hpp>
 #include <SFML/System/Err.hpp>
 #include <X11/Xlib.h>
 #include <X11/extensions/Xrandr.h>
